requirements.txt
# Core
httpx>=0.25.0
pydantic>=2.0.0
sqlite-utils>=3.35.0
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
typer>=0.9.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
respx>=0.21.0
